+++ src/main/java/module-info.java
@@ -4,6 +4,6 @@ module sdu.se9.tv2.management.system {
     requires kotlin.stdlib;
     requires json.simple;
 
-    opens sdu.se9.tv2.management.system to javafx.fxml;
-    exports sdu.se9.tv2.management.system;
+    opens sdu.se9.tv2.management.system.presentation to javafx.fxml;
+    exports sdu.se9.tv2.management.system.presentation;
 }
