+++ src/main/java/sdu/se9/tv2/management/system/domain/Producer.java
@@ -1,4 +1,4 @@
-package sdu.se9.tv2.management.system;
+package sdu.se9.tv2.management.system.domain;
 
 import org.json.simple.JSONObject;
