+++ src/main/java/sdu/se9/tv2/management/system/domain/accounts/AdminAccount.java
@@ -1,7 +1,7 @@
-package sdu.se9.tv2.management.system;
+package sdu.se9.tv2.management.system.domain.accounts;
 
 public class AdminAccount extends Account {
-    AdminAccount(int accountId, String accountUsername, String accountPassword) {
+    public AdminAccount(int accountId, String accountUsername, String accountPassword) {
         super(accountId, accountUsername, accountPassword,  "admin");
     }
