+++ src/main/java/sdu/se9/tv2/management/system/domain/accounts/ProducerAccount.java
@@ -1,9 +1,9 @@
-package sdu.se9.tv2.management.system;
+package sdu.se9.tv2.management.system.domain.accounts;
 
 public class ProducerAccount extends Account {
     private int producerId;
 
-    ProducerAccount(int accountId, String accountUsername, String accountPassword, int producerId) {
+    public ProducerAccount(int accountId, String accountUsername, String accountPassword, int producerId) {
         super(accountId, accountUsername, accountPassword, "producer");
 
         this.producerId = producerId;
