+++ src/main/java/sdu/se9/tv2/management/system/persistence/IPersistenceAccount.java
@@ -1,5 +1,9 @@
-package sdu.se9.tv2.management.system;
+package sdu.se9.tv2.management.system.persistence;
 
+import sdu.se9.tv2.management.system.domain.accounts.Account;
+import sdu.se9.tv2.management.system.domain.accounts.AdminAccount;
+import sdu.se9.tv2.management.system.domain.accounts.ProducerAccount;
+import sdu.se9.tv2.management.system.domain.accounts.SystemAdminAccount;
 import sdu.se9.tv2.management.system.exceptions.UsernameAlreadyExistsException;
 
 /**
