+++ src/main/java/sdu/se9/tv2/management/system/persistence/PersistenceAccount.java
@@ -1,8 +1,12 @@
-package sdu.se9.tv2.management.system;
+package sdu.se9.tv2.management.system.persistence;
 
 import org.json.simple.JSONArray;
 import org.json.simple.JSONObject;
 import org.json.simple.parser.ParseException;
+import sdu.se9.tv2.management.system.domain.accounts.Account;
+import sdu.se9.tv2.management.system.domain.accounts.AdminAccount;
+import sdu.se9.tv2.management.system.domain.accounts.ProducerAccount;
+import sdu.se9.tv2.management.system.domain.accounts.SystemAdminAccount;
 import sdu.se9.tv2.management.system.exceptions.UsernameAlreadyExistsException;
 
 import java.io.InvalidClassException;
