+++ src/main/java/sdu/se9/tv2/management/system/persistence/PersistencePerson.java
@@ -1,10 +1,11 @@
-package sdu.se9.tv2.management.system;
+package sdu.se9.tv2.management.system.persistence;
 import org.json.simple.parser.ParseException;
 import java.util.ArrayList;
 import java.util.Iterator;
 
 import org.json.simple.JSONArray;
 import org.json.simple.JSONObject;
+import sdu.se9.tv2.management.system.domain.Person;
 
 /**
  * Implementation of the IPersistencePerson interface
@@ -28,7 +29,7 @@ public class PersistencePerson implements IPersistencePerson {
     /**
      * Creates a new instance of the PersistencePerson class
      */
-    PersistencePerson () {
+    public PersistencePerson() {
         // Once a new instance is made the data will be read and saved in memory
         this.read();
     }
