+++ src/main/java/sdu/se9/tv2/management/system/presentation/PrimaryController.java
@@ -1,7 +1,8 @@
-package sdu.se9.tv2.management.system;
+package sdu.se9.tv2.management.system.presentation;
 
 import java.io.IOException;
 import javafx.fxml.FXML;
+import sdu.se9.tv2.management.system.presentation.App;
 
 public class PrimaryController {
