+++ src/main/resources/sdu/se9/tv2/management/system/secondary.fxml
@@ -5,7 +5,7 @@
 <?import javafx.scene.control.Button?>
 <?import javafx.geometry.Insets?>
 
-<VBox alignment="CENTER" spacing="20.0" xmlns="http://javafx.com/javafx/8.0.171" xmlns:fx="http://javafx.com/fxml/1" fx:controller="sdu.se9.tv2.management.system.SecondaryController">
+<VBox alignment="CENTER" spacing="20.0" xmlns="http://javafx.com/javafx/8.0.171" xmlns:fx="http://javafx.com/fxml/1" fx:controller="sdu.se9.tv2.management.system.presentation.SecondaryController">
     <children>
         <Label text="Secondary View" />
         <Button fx:id="secondaryButton" text="Switch to Primary View" onAction="#switchToPrimary" />
